Develop

+++ quickOS.Application/DTOs/OutputModels/ServiceProvidedOutputModel.cs
@@ -6,10 +6,10 @@ public class ServiceProvidedOutputModel
     public int Code { get; private set; }
     public string Name { get; private set; }
     public string Description { get; private set; }
-    public double Price { get; private set; }
+    public decimal Price { get; private set; }
     public bool IsActive { get; private set; }
 
-    public ServiceProvidedOutputModel(Guid externalId, int code, string name, string description, double price, bool isActive)
+    public ServiceProvidedOutputModel(Guid externalId, int code, string name, string description, decimal price, bool isActive)
     {
         ExternalId = externalId;
         Code = code;
